tasks/generate-assessment-chart-configs/requirements.txt
functions-framework>=3.10.1
google-cloud-bigquery>=3.34.0
google-cloud-storage>=3.9.0
